chore: eliminate circular dependencies and add madge for dependency analysis
This commit improves the internal code structure by eliminating all circular dependencies and adding tooling to prevent future issues.

Changes:
- Introduced madge for circular dependency detection and visualization
- Eliminated circular dependencies by merging type definitions from utils/types.ts into common/types.ts
- Fixed import paths in source files to use common/types.ts
- Added npm scripts for dependency analysis (check:circular, graph:*, etc.)
- Added circular dependency check to CI pipeline
- Updated .gitignore to exclude generated dependency graph files
