+++ .github/workflows/.build.yaml
@@ -19,6 +19,8 @@ jobs:
         run: pnpm run build
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      - name: Check Circular Dependencies
+        run: pnpm check:circular
       - name: Type Check
         run: pnpm check:type
       - uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
