+++ package.json
@@ -113,8 +113,16 @@
     "bench:rust": "cargo bench --manifest-path web-csv-toolbox-wasm/Cargo.toml",
     "coverage:rust": "cargo llvm-cov --manifest-path web-csv-toolbox-wasm/Cargo.toml --lcov --output-path lcov-rust.info",
     "check:type": "tsc --noEmit",
+    "check:circular": "madge --circular --extensions ts src/",
     "check:format": "pnpm check:format:js",
     "check:format:js": "biome ci .",
+    "graph": "madge --extensions ts --image graph.svg src/",
+    "graph:main": "madge --extensions ts --image graph-main.svg src/web-csv-toolbox.ts",
+    "graph:worker": "madge --extensions ts --image graph-worker.svg src/worker.node.ts src/worker.web.ts",
+    "graph:json": "madge --extensions ts --json src/",
+    "graph:summary": "madge --extensions ts --summary src/",
+    "graph:orphans": "madge --extensions ts --orphans src/",
+    "graph:leaves": "madge --extensions ts --leaves src/",
     "test:browser": "vitest --browser --project browser",
     "test:coverage": "vitest --coverage",
     "test:bench": "pnpm --filter web-csv-toolbox-benchmark start",
@@ -157,6 +165,7 @@
     "@wasm-tool/rollup-plugin-rust": "^3.0.5",
     "changesets-github-release": "^0.1.0",
     "fast-check": "^4.1.1",
+    "madge": "^8.0.0",
     "terser": "^5.42.0",
     "typedoc": "^0.28.5",
     "typedoc-plugin-mdn-links": "^5.0.2",
