+++ src/parse.ts
@@ -5,6 +5,7 @@ import type {
   CSVString,
   ParseBinaryOptions,
   ParseOptions,
+  PickCSVHeader,
 } from "./common/types.ts";
 import type { DEFAULT_DELIMITER, DEFAULT_QUOTATION } from "./constants.ts";
 import { parseBinary } from "./parseBinary.ts";
@@ -15,7 +16,6 @@ import { parseString } from "./parseString.ts";
 import { parseStringStream } from "./parseStringStream.ts";
 import { parseUint8ArrayStream } from "./parseUint8ArrayStream.ts";
 import * as internal from "./utils/convertThisAsyncIterableIteratorToArray.ts";
-import type { PickCSVHeader } from "./utils/types.ts";
 
 /**
  * Parse CSV to records.
