+++ src/parseStringToArraySyncWASM.ts
@@ -1,13 +1,16 @@
 import { parseStringToArraySync } from "web-csv-toolbox-wasm";
 import { assertCommonOptions } from "./assertCommonOptions.ts";
-import type { CSVRecord, CommonOptions } from "./common/types.ts";
+import type {
+  CSVRecord,
+  CommonOptions,
+  PickCSVHeader,
+} from "./common/types.ts";
 import {
   DEFAULT_DELIMITER,
   DEFAULT_QUOTATION,
   DOUBLE_QUOTE,
 } from "./constants.ts";
 import type { loadWASM } from "./loadWASM.ts";
-import type { PickCSVHeader } from "./utils/types.ts";
 
 /**
  * Parse CSV string to record of arrays.
